refactor(renaming): Move TypeScript renaming into its own class
